+++ src/Query/Condition/ConditionsInterface.php
@@ -6,4 +6,9 @@ namespace WoohooLabs\Larva\Query\Condition;
 interface ConditionsInterface
 {
     public function getConditions(): array;
+
+    /**
+     * @return void
+     */
+    public function addConditionGroup(ConditionBuilderInterface $conditions, string $operator = "AND");
 }